update: Added 2017 TCs

+++ scripts/update.py
@@ -1,10 +1,12 @@
-import shutil
+from pathlib import Path
+from re import findall
+from shutil import copyfile
 from zipfile import ZipFile
 from glob import glob
-from StringIO import StringIO
+from io import StringIO
 import pandas as pd
 
-IN_FILE = 'output/JTWC_raw.csv'
+IN_FILE = 'input/JTWC_raw.csv'
 BAK_FILE = 'output/JTWC_raw.csv.bak'
 IN_DIR = 'input/raw/'
 
@@ -69,17 +71,28 @@ def parse_input(filepath_or_buffer):
 
 
 zip_files = glob(IN_DIR + '*.zip')
+zip_files = pd.DataFrame(zip_files, index=[int(findall('[0-9]{4}',z)[0]) for z in zip_files], columns=['name'])
+
+in_file = Path(IN_FILE)
+if in_file.exists():
+    in_df = pd.read_csv(in_file)
+    zip_files = zip_files[~zip_files.index.isin(in_df['YYYY'])]
+else:
+    in_df = pd.DataFrame()
 
 df = None
-for zfile in zip_files:
+for zfile in zip_files['name']:
     with ZipFile(zfile) as z:
         for f in z.namelist():
-            ff = StringIO(z.read(f).replace(' ', ''))
+            ff = StringIO(z.read(f).decode('utf-8').replace(' ', ''))
             if df is None:
                 df = parse_input(ff)
             df = pd.concat([df, parse_input(ff)])
 
-out_df = df.sort_values(['YY', 'MM', 'DD', 'HH', 'CY'])
+out_df = df.sort_values(['YYYY', 'MM', 'DD', 'HH', 'CY'])
 
-shutil.copyfile(IN_FILE, BAK_FILE)  # Backup the file
-out_df.to_csv(IN_FILE, mode='a', index=False, header=False)
+if in_file.exists():
+    copyfile(IN_FILE, BAK_FILE)  # Backup the file
+    out_df.to_csv(IN_FILE, mode='a', index=False, header=False)
+else:
+    out_df.to_csv(IN_FILE, index=False)
